Tracer function incomplete (Animation times)
Each hive block is able to assign them self a proper number in the circuit and correct on update when ever a new block is placed directly next to them. Animation times do not sync up yet and may need to implement method to update other than just on collision.